pyproject.toml
[project]
name = "stats-code"
version = "0.1.2"
description = "A full python tool to gather statistics from your codebase."
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.11"
dependencies = [
    "chardet>=5.2.0",
    "pathspec>=0.12.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
]
authors = [
    { name = "LKLLL", email = "13263311018@163.com" }
]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
keywords = ["code", "developer", "statistics", "stats-code"]
urls = { "Home" = "https://github.com/LKLLLLLLLLLL/CodeStatistics"}

[tool.taskipy.tasks]
update-config = "python -m scripts.update_config"
build = "python -m scripts.build"

[project.scripts]
stats-code = "stats_code.__main__:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["stats_code"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
stats_code = ["config/*.yml"]

[dependency-groups]
dev = [
    "jinja2>=3.1.6",
    "pydantic>=2.12.3",
    "taskipy>=1.14.1",
    "twine>=6.2.0",
    "viztracer>=1.1.0",
]
